Re-factored the project to use the feature pattern.
I did this because I have been a part of enough projects (small and large)
that I have seen the benefits of beginning with the feature pattern from
the start of a project. Yes you do have more files referenced to in your
index.html file. But that is what projects like use min is for:
https://www.npmjs.org/package/gulp-usemin